requirements.txt
pyyaml
aiohttp
python-frontmatter
transformers
pydantic
pydantic-core
numpy
pandas
haystack-ai
milvus-haystack
pymilvus
sentence-transformers
fastapi
uvicorn[standard]
pydantic_settings
frontmatter
nltk
python-multipart
celery